#15 Allow DDL-statements to be executed asynchronously
A connection can be set to execute DDL-operations in async mode. When a
DDL-statement is executed asynchronously, it is submitted to Google
Cloud Spanner and checked for syntax and other basic errors. The
statement then returns while Cloud Spanner executes the statement in the
background. The status of the execution can be queried by issuing the
statement 'SHOW_DDL_OPERATIONS'.