allow resetting a dynamic connection property
By allowing a dynamic connection property to be reset, it is easier to
set and reset these properties in scripts. For example, when running a
script containing a number of DDL-statements, the script can temporarily
set the property AsyncDdlOperations to true, and then resetting it to
its original value after the statements have been executed.