#51 put a limit on the count query to determine extended mode
The count query that determines whether an statement should be executed
in extended mode does not need to know the exact number of records, only
if the number of records is at least the threshold for entering extended
mode. Putting a limit on this query equal to the threshold speeds up the
query and prevents an exception (out of memory) from occurring in Cloud
Spanner on very big tables.