transaction thread did not perform a real rollback
A transaction that was rolled back did not actually perform a rollback,
but instead threw away all buffered mutations and did an empty commit.
While functionally equivalent, it is better to do a real rollback in
case Spanner in the future might actually support writing mutations
without also committing the transaction at the same time, which would
require the driver to do an actual rollback.