#64 added tests for AbortedException
Spanner can abort transactions in case of detected deadlocks or other
problems, but this is hard to debug as it is hard to actually trigger
this problem on spanner. The emulator can trigger an aborted transaction
on demand. This test shows that an aborted transaction during commit
will be retried successfully. Aborted transactions that are used for
reading, will still fail. It used to fail with a SpannerException, it
now fails with an SQLException.