#77 isWritable and related methods returned false for pk columns
isWritable, isDefinitelyWritable and isReadOnly considered primary key
columns as read-only. This is not really true, as primary key columns
are writable during an insert, but not updatable later. The columns
themselves are therefore writable (they are not computed or derived
columns), but the value cannot be changed once it has been written.